--- a/tests/NewId.Tests/Generator_Specs.cs
+++ b/tests/NewId.Tests/Generator_Specs.cs
@@ -1,6 +1,4 @@
-﻿using System.Linq;
-
-namespace MassTransit.NewIdTests
+﻿namespace MassTransit.NewIdTests
 {
     using System;
     using System.Diagnostics;
@@ -93,7 +91,6 @@ namespace MassTransit.NewIdTests
 
             NewId nid1 = id1.ToNewId();
             NewId nid2 = id2.ToNewId();
-
         }
 
         [Test]
@@ -118,6 +115,40 @@ namespace MassTransit.NewIdTests
             NewId nid2 = id2.ToNewIdFromSequential();
 
             Assert.AreEqual(nid, nid1);
+        }
+
+        [Test]
+        public void Should_generate_known_sequential_guid()
+        {
+            var expected = Guid.Parse("74b719ec-7596-3cf0-7d81-bf34437f0b01");
+            var tickProvider = new MockTickProvider(8410219332513447152);
+            var networkProvider = new MockNetworkProvider(BitConverter.GetBytes(6857996259202924925));
+            var generator = new NewIdGenerator(tickProvider, networkProvider);
+
+            for (int i = 0; i < 267; i++)
+            {
+                generator.NextSequentialGuid();
+            }
+            var guid = generator.NextSequentialGuid();
+
+            Assert.AreEqual(expected, guid);
+        }
+
+        [Test]
+        public void Should_generate_known_guid()
+        {
+            var expected = Guid.Parse("437f0b01-bf34-7d81-3cf0-74b719ec7596");
+            var tickProvider = new MockTickProvider(8410219332513447152);
+            var networkProvider = new MockNetworkProvider(BitConverter.GetBytes(6857996259202924925));
+            var generator = new NewIdGenerator(tickProvider, networkProvider);
+
+            for (int i = 0; i < 267; i++)
+            {
+                generator.NextGuid();
+            }
+            var guid = generator.NextGuid();
+
+            Assert.AreEqual(expected, guid);
         }
 
         [SetUp]
@@ -170,7 +201,6 @@ namespace MassTransit.NewIdTests
                 return _workerId;
             }
         }
-
 
         class MockProcessIdProvider :
             IProcessIdProvider