I've made some changes to NextGuid and NextSequentialGuid to improve their performance. These changes include vectorizing the functions, which has resulted in a speed improvement of 10-15%. As NextGuid is frequently called in MassTransit, this small improvement should add up.

Also added some simple assertion tests to ensure that the changes do not break any existing functionality. Additionally, I've made some minor code improvements.

#### Benchmarks
Benchmark code can be [found here](https://github.com/TimothyMakkison/NewId/blob/vectorize_guid_benchmarks/tests/NewId.Benchmarks/NewIdBenchmarks.cs).

|                     Method | EnvironmentVariables |              Mean |             Error |            StdDev | Ratio |  Allocated |
|--------------------------- |--------------------- |------------------:|------------------:|------------------:|------:|-----------:|
|                   NextGuid | COMPlus_EnableSSE2=0 |          63.68 ns |          0.485 ns |          0.430 ns |  1.00 |          - |
|                   NextGuid |                Empty |          53.90 ns |          0.124 ns |          0.116 ns |  0.85 |          - |
|                            |                      |                   |                   |                   |       |            |
|               NextGuidBulk | COMPlus_EnableSSE2=0 |   6,499,998.33 ns |     37,311.406 ns |     33,075.590 ns |  1.00 |        5 B |
|               NextGuidBulk |                Empty |   5,517,873.10 ns |     11,506.953 ns |     10,200.614 ns |  0.85 |        5 B |
|                            |                      |                   |                   |                   |       |            |
|         NextSequentialGuid | COMPlus_EnableSSE2=0 |          62.39 ns |          0.111 ns |          0.098 ns |  1.00 |          - |
|         NextSequentialGuid |                Empty |          53.58 ns |          0.155 ns |          0.129 ns |  0.86 |          - |
|                            |                      |                   |                   |                   |       |            |
|     NextSequentialGuidBulk | COMPlus_EnableSSE2=0 |   6,379,690.66 ns |      9,309.096 ns |      7,773.517 ns |  1.00 |        5 B |
|     NextSequentialGuidBulk |                Empty |   5,520,476.59 ns |      9,472.339 ns |      8,860.432 ns |  0.87 |        5 B |
|                            |                      |                   |                   |                   |       |            |
|           NextGuidParallel | COMPlus_EnableSSE2=0 | 372,620,013.00 ns | 12,028,605.371 ns | 35,466,614.329 ns |  1.00 | 83891776 B |
|           NextGuidParallel |                Empty | 311,851,534.42 ns |  6,214,288.576 ns | 15,929,597.331 ns |  0.86 | 83891312 B |
|                            |                      |                   |                   |                   |       |            |
| NextSequentialGuidParallel | COMPlus_EnableSSE2=0 | 349,581,895.92 ns | 10,806,814.204 ns | 31,523,984.276 ns |  1.00 | 83891776 B |
| NextSequentialGuidParallel |                Empty | 324,188,729.55 ns |  6,403,590.216 ns | 15,094,015.521 ns |  0.91 | 83891312 B |


I've also experimented with a non-intrinsic version of NextGuid and NextSequentialGuid that achieves a 5-10% speed improvement, but the results are very unreliable. If anyone is interested in trying to improve upon this version, you can find the [code here](https://github.com/TimothyMakkison/NewId/blob/52aec9a17a0db87e8ef19f4c264332b36eb4f9f3/src/NewId/NewIdGenerator.cs#L105)