@TimothyMakkison 2023-01-04
Thanks I'll make a pr on MassTransit, I'll try to optimize `NextGuid(Guid[] ids, int index, int count)`